move EntityCrossProjectSelectionConfirmModal back to index.ts for use in AssayImportSubMenuItem
