+++ packages/components/src/entities/CreateSamplesSubMenuBase.tsx
@@ -12,7 +12,7 @@ import { SCHEMAS } from '../internal/schemas';
 
 import { getCrossFolderSelectionResult } from '../internal/components/entities/actions';
 
-import { EntityCrossProjectSelectionConfirmModal } from './EntityCrossProjectSelectionConfirmModal';
+import { EntityCrossProjectSelectionConfirmModal } from '../internal/components/entities/EntityCrossProjectSelectionConfirmModal';
 
 import { isSamplesSchema } from '../internal/components/samples/utils';
 import { SampleCreationTypeModal } from './SampleCreationTypeModal';
