+++ packages/components/src/entities/index.ts
@@ -26,7 +26,6 @@ import { CreateSamplesSubMenu } from './CreateSamplesSubMenu';
 import { CreateSamplesSubMenuBase } from './CreateSamplesSubMenuBase';
 import { SamplesAssayButton } from './SamplesAssayButton';
 import { SampleTypeInsightsPanel } from './SampleTypeInsightsPanel';
-import { EntityCrossProjectSelectionConfirmModal } from './EntityCrossProjectSelectionConfirmModal';
 import { FindSamplesByIdHeaderPanel } from './FindSamplesByIdHeaderPanel';
 import { FindSamplesByIdsPageBase } from './FindSamplesByIdsPageBase';
 import { SampleFinderSection } from './SampleFinderSection';
@@ -62,7 +61,6 @@ export {
     loadSampleTypes,
     CreateSamplesSubMenu,
     CreateSamplesSubMenuBase,
-    EntityCrossProjectSelectionConfirmModal,
     EntityDeleteModal,
     EntityLineageEditMenuItem,
     EntityTypeDeleteConfirmModal,
