+++ packages/components/src/index.ts
@@ -441,6 +441,7 @@ import {
 } from './internal/components/entities/constants';
 import { getUniqueIdColumnMetadata } from './internal/components/entities/utils';
 import { EntityInsertPanel } from './internal/components/entities/EntityInsertPanel';
+import { EntityCrossProjectSelectionConfirmModal } from './internal/components/entities/EntityCrossProjectSelectionConfirmModal';
 import { SampleTypeModel } from './internal/components/domainproperties/samples/models';
 
 import { EditableDetailPanel } from './public/QueryModel/EditableDetailPanel';
@@ -1038,6 +1039,7 @@ export {
     getSelectionLineageData,
     getLineageEditorUpdateColumns,
     EntityInsertPanel,
+    EntityCrossProjectSelectionConfirmModal,
     SampleTypeDataType,
     DataClassDataType,
     AssayResultDataType,
